Main application loop v1.0
Created run method with basic multithreading capabilities (no tests for it actually
working, it just compiles and runs as expected at the moment).

Working mechanics:

execution_context_ is running in some number of
threads defined by THREADS_COUNT macro (this way one execution_context_
can split tasks assigned to it between multiple threads).

execution_context_ has work_guard_ applied to it, so it
doesn't return if there is no work. So at the moment of this commit if
you were to compile this you would see that all threads are just waiting.

Misc:

1. Added logger class

2. Project now has Makefile (soon I will update README with quick guide)

3. Some precompiled Boost libs now included in CMake with find_package()